# Tests: Use "UNION ALL" instead of "UNION" to improve performance