--- a/test/acceptance/turbo-carto/named-maps.js
+++ b/test/acceptance/turbo-carto/named-maps.js
@@ -45,13 +45,13 @@ describe('turbo-carto for named maps', function() {
                             sql: [
                                 'SELECT ' + table + '.*, _prices.price FROM ' + table + ' JOIN (' +
                                 '  SELECT 1 AS cartodb_id, 10.00 AS price',
-                                '  UNION',
+                                '  UNION ALL',
                                 '  SELECT 2, 10.50',
-                                '  UNION',
+                                '  UNION ALL',
                                 '  SELECT 3, 11.00',
-                                '  UNION',
+                                '  UNION ALL',
                                 '  SELECT 4, 12.00',
-                                '  UNION',
+                                '  UNION ALL',
                                 '  SELECT 5, 21.00',
                                 ') _prices ON _prices.cartodb_id = ' + table + '.cartodb_id'
                             ].join('\n'),