--- a/test/acceptance/widgets/regressions.js
+++ b/test/acceptance/widgets/regressions.js
@@ -265,12 +265,12 @@ describe('widgets-regressions', function() {
                     ST_TRANSFORM(ST_SETSRID(ST_GeomFromGeoJSON(
                         '${JSON.stringify(notIntersectingLeftTriangle)}'
                     ), 4326), 3857) AS the_geom_webmercator, 1 AS cartodb_id, 'notIntersectingLeftTriangle' AS name
-                UNION
+                UNION ALL
                 SELECT
                     ST_TRANSFORM(ST_SETSRID(ST_GeomFromGeoJSON(
                         '${JSON.stringify(notIntersectingRightTriangle)}'
                     ), 4326), 3857), 2, 'notIntersectingRightTriangle'
-                UNION
+                UNION ALL
                 SELECT
                     ST_TRANSFORM(ST_SETSRID(ST_GeomFromGeoJSON(
                         '${JSON.stringify(intersectingTriangle)}'