While investigating https://github.com/CartoDB/Windshaft-cartodb/pull/1002 with @jgoizueta I noticed that when combining queries using `UNION` a query would take ~60ms as it was generating a plan expecting both sides of the union being 1 million rows in size.

On the other hand, using `UNION ALL` would take ~1 ms, using a different plan as it doesn't require to remove duplicates from the combination. @jgoizueta tested with PG 9.6 and it varied (60ms the first time, and ~1ms the rest) and in production (PG 10.1) where it was always 1ms. My installation (PG11) always went to the 60 ms plan.

As in these cases there is no need to remove duplicates, UNION ALL is more efficient.